Move machine removals into their own file in state.
